@@ -1,21 +1,77 @@
+/**
+ * Flugbókanakerfi HBV401G Hópur 1F
+ * Alda, Ármann, Halldór og Hrólfur
+ */
+
 package is.hi.tripplanner.vinnsla;
 
-import java.util.Random;
+import java.time.LocalDateTime;
 
 public class Flight {
-    private long flightID;
-    private String Departure;
+    private String FlightID;
+    private String departureLoc;
     private String destination;
-    private int arrivalDate;
-    private int departureDate;
-    private int price;
+    private LocalDateTime departureTime;
+    private LocalDateTime arrivalTime;
     private int availableSeats;
+    private float price;
 
-    public Flight() {
-        this.flightID = new Random().nextLong();
+    public Flight(String FlightID, String departureLoc, String destination,
+                  LocalDateTime departureTime, LocalDateTime arrivalTime,
+                  int availableSeats, float price){
+        this.FlightID = FlightID;
+        this.departureLoc = departureLoc;
+        this.destination = destination;
+        this.departureTime = departureTime;
+        this.arrivalTime = arrivalTime;
+        this.availableSeats = availableSeats;
+        this.price = price;
     }
 
-    public long getFlightID() {
-        return this.flightID;
+    public String getFlightID() {
+        return FlightID;
+    }
+
+    public String getDepartureLoc() {
+        return departureLoc;
+    }
+
+    public String getDestination() {
+        return destination;
+    }
+
+    public LocalDateTime getDepartureTime() {
+        return departureTime;
+    }
+
+    public LocalDateTime getArrivalTime() {
+        return arrivalTime;
+    }
+
+    public int getAvailableSeats() {
+        return availableSeats;
+    }
+
+    public float getPrice() {
+        return price;
+    }
+    /*
+    public String toString() {
+        return FlightID + " " + departureLoc + " " + destination + " " + departureTime + " " + arrivalTime +
+                " " + availableSeats + " " + price + " ";
+    }
+     */
+
+    @Override
+    public String toString() {
+        return "Flight item:\n" +
+                "FlightID='" + FlightID + '\'' +
+                ", departureLoc='" + departureLoc + '\'' +
+                ", destination='" + destination + '\'' +
+                ", departureTime=" + departureTime +
+                ",\n arrivalTime=" + arrivalTime +
+                ", availableSeats=" + availableSeats +
+                ", price=" + price +
+                "\n";
     }
 }
